auth-proxy/requirements.txt
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0; platform_python_implementation == "CPython" and python_version < "3.9"
